Refactor code, split servers, move things around
